+++ src/controllers/connector/imweb/ImwebController.ts
@@ -48,11 +48,8 @@ export class ImwebController {
    * to get the detailed information of a sale, use {@link at} function for
    * each sale.
    *
-   * For reference, if you're a {@link IShoppingSeller seller}, you can only
-   * access to the your own {@link IShoppingSale sale}s. Otherwise you're a
-   * {@link IShoppingCustomer customer}, you can see only the operating
-   * sales in the market. Instead, you can't see the unopened, closed, or
-   * suspended sales.
+   * you can see only the operating sales in the market.
+   * Instead, you can't see the unopened, closed, or suspended sales.
    *
    * By the way, if you want, you can limit the result by configuring
    * {@link IShoppingSale.IRequest.search search condition} in the request
